Move POSIX specific startup to dedicated build file
